拉取请求

+++ redisson-tomcat/redisson-tomcat-9/src/main/java/org/redisson/tomcat/RedissonSession.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2013-2019 Nikita Koksharov
+ * Copyright (c) 2013-2020 Nikita Koksharov
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -29,6 +29,7 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 
 import org.apache.catalina.session.StandardSession;
+import org.redisson.api.RSet;
 import org.redisson.api.RMap;
 import org.redisson.api.RTopic;
 import org.redisson.tomcat.RedissonSessionManager.ReadMode;
@@ -48,10 +49,15 @@ public class RedissonSession extends StandardSession {
     private static final String MAX_INACTIVE_INTERVAL_ATTR = "session:maxInactiveInterval";
     private static final String LAST_ACCESSED_TIME_ATTR = "session:lastAccessedTime";
     private static final String CREATION_TIME_ATTR = "session:creationTime";
+    private static final String IS_EXPIRATION_LOCKED = "session:isExpirationLocked";
     
-    public static final Set<String> ATTRS = new HashSet<String>(Arrays.asList(IS_NEW_ATTR, IS_VALID_ATTR, 
-            THIS_ACCESSED_TIME_ATTR, MAX_INACTIVE_INTERVAL_ATTR, LAST_ACCESSED_TIME_ATTR, CREATION_TIME_ATTR));
+    public static final Set<String> ATTRS = new HashSet<String>(Arrays.asList(
+            IS_NEW_ATTR, IS_VALID_ATTR, 
+            THIS_ACCESSED_TIME_ATTR, MAX_INACTIVE_INTERVAL_ATTR, 
+            LAST_ACCESSED_TIME_ATTR, CREATION_TIME_ATTR, IS_EXPIRATION_LOCKED
+            ));
     
+    private boolean isExpirationLocked;
     private boolean loaded;
     private final RedissonSessionManager redissonManager;
     private final Map<String, Object> attrs;
@@ -61,16 +67,21 @@ public class RedissonSession extends StandardSession {
     private final UpdateMode updateMode;
     
     private Set<String> removedAttributes = Collections.emptySet();
+    private Set<String> updatedAttributes = Collections.emptySet();
     
-    public RedissonSession(RedissonSessionManager manager, RedissonSessionManager.ReadMode readMode, UpdateMode updateMode) {
+    private final boolean broadcastSessionEvents;
+    
+    public RedissonSession(RedissonSessionManager manager, RedissonSessionManager.ReadMode readMode, UpdateMode updateMode, boolean broadcastSessionEvents) {
         super(manager);
         this.redissonManager = manager;
         this.readMode = readMode;
         this.updateMode = updateMode;
         this.topic = redissonManager.getTopic();
+        this.broadcastSessionEvents = broadcastSessionEvents;
         
         if (updateMode == UpdateMode.AFTER_REQUEST) {
             removedAttributes = Collections.newSetFromMap(new ConcurrentHashMap<String, Boolean>());
+            updatedAttributes = Collections.newSetFromMap(new ConcurrentHashMap<String, Boolean>());
         }
         
         try {
@@ -92,6 +103,11 @@ public class RedissonSession extends StandardSession {
 
             if (name == null) {
                 return null;
+            }
+
+            if (updatedAttributes.contains(name)
+                    || removedAttributes.contains(name)) {
+                return super.getAttribute(name);
             }
 
             return map.get(name);
@@ -143,7 +159,15 @@ public class RedissonSession extends StandardSession {
             map = redissonManager.getMap(id);
         }
         
-        map.delete();
+        if (broadcastSessionEvents) {
+            RSet<String> set = redissonManager.getNotifiedNodes(id);
+            set.add(redissonManager.getNodeId());
+            set.expire(60, TimeUnit.SECONDS);
+            map.fastPut(IS_EXPIRATION_LOCKED, true);
+            map.expire(60, TimeUnit.SECONDS);
+        } else {
+            map.delete();
+        }
         if (readMode == ReadMode.MEMORY) {
             topic.publish(new AttributesClearMessage(redissonManager.getNodeId(), getId()));
         }
@@ -183,11 +207,14 @@ public class RedissonSession extends StandardSession {
     }
 
     protected void expireSession() {
+        if (isExpirationLocked) {
+            return;
+        }
         if (maxInactiveInterval >= 0) {
             map.expire(maxInactiveInterval + 60, TimeUnit.SECONDS);
         }
     }
-    
+
     protected AttributesPutAllMessage createPutAllMessage(Map<String, Object> newMap) {
         Map<String, Object> map = new HashMap<String, Object>();
         for (Entry<String, Object> entry : newMap.entrySet()) {
@@ -261,11 +288,15 @@ public class RedissonSession extends StandardSession {
     public void setAttribute(String name, Object value, boolean notify) {
         super.setAttribute(name, value, notify);
         
-        if (updateMode == UpdateMode.DEFAULT && map != null && value != null) {
+        if (value == null) {
+            return;
+        }
+        if (updateMode == UpdateMode.DEFAULT && map != null) {
             fastPut(name, value);
         }
         if (updateMode == UpdateMode.AFTER_REQUEST) {
             removedAttributes.remove(name);
+            updatedAttributes.add(name);
         }
     }
     
@@ -285,6 +316,7 @@ public class RedissonSession extends StandardSession {
         }
         if (updateMode == UpdateMode.AFTER_REQUEST) {
             removedAttributes.add(name);
+            updatedAttributes.remove(name);
         }
     }
     
@@ -300,6 +332,9 @@ public class RedissonSession extends StandardSession {
         newMap.put(MAX_INACTIVE_INTERVAL_ATTR, maxInactiveInterval);
         newMap.put(IS_VALID_ATTR, isValid);
         newMap.put(IS_NEW_ATTR, isNew);
+        if (broadcastSessionEvents) {
+            newMap.put(IS_EXPIRATION_LOCKED, isExpirationLocked);
+        }
         
         if (attrs != null) {
             for (Entry<String, Object> entry : attrs.entrySet()) {
@@ -319,6 +354,8 @@ public class RedissonSession extends StandardSession {
                 }
             }
         }
+
+        updatedAttributes.clear();
         removedAttributes.clear();
         
         expireSession();
@@ -348,6 +385,10 @@ public class RedissonSession extends StandardSession {
         Boolean isNew = (Boolean) attrs.remove(IS_NEW_ATTR);
         if (isNew != null) {
             this.isNew = isNew;
+        }
+        Boolean isExpirationLocked = (Boolean) attrs.remove(IS_EXPIRATION_LOCKED);
+        if (isExpirationLocked != null) {
+            this.isExpirationLocked = isExpirationLocked;
         }
 
         for (Entry<String, Object> entry : attrs.entrySet()) {
